compile hdg toolboxes as libraries
@lsala @prudhomm
