+++ toolboxes/feel/feelmodels/hdg/mixedelasticityassembly_inst.cpp
@@ -1,0 +1,11 @@
+
+#include "mixedelasticityconfig.h"
+#include <feel/feelmodels/hdg/mixedelasticityassembly.cpp>
+
+namespace Feel {
+namespace FeelModels {
+
+MIXEDELASTICITY_CLASS_INSTANTIATION
+
+}
+}
