+++ toolboxes/feel/feelmodels/hdg/mixedelasticityconfig.h.in
@@ -1,0 +1,20 @@
+#undef MIXEDELASTICITY_DIM
+#undef MIXEDELASTICITY_ORDERGEO
+#undef MIXEDELASTICITY_ORDERPOLY
+
+#define MIXEDELASTICITY_DIM @MIXEDELASTICITY_DIM@
+#define MIXEDELASTICITY_ORDERGEO @MIXEDELASTICITY_ORDERGEO@
+#define MIXEDELASTICITY_ORDERPOLY @MIXEDELASTICITY_ORDERPOLY@
+
+#define MIXEDELASTICITY_CLASS_TEMPLATE_DECLARATIONS \
+    template<int Dim, int Order, int G_Order, int E_Order> \
+        /**/
+#define MIXEDELASTICITY_CLASS_TEMPLATE_TYPE \
+    MixedElasticity<Dim, Order, G_Order, E_Order> \
+        /**/
+#define MIXEDELASTICITY_CLASS_TYPE \
+    MixedElasticity< MIXEDELASTICITY_DIM,MIXEDELASTICITY_ORDERPOLY,MIXEDELASTICITY_ORDERGEO> \
+    /**/
+#define MIXEDELASTICITY_CLASS_INSTANTIATION \
+    template class MIXEDELASTICITY_CLASS_TYPE ;  \
+    /**/
