+++ toolboxes/feel/feelmodels/hdg/mixedelasticitycreate_inst.cpp
@@ -1,0 +1,11 @@
+
+#include "mixedelasticityconfig.h"
+#include <feel/feelmodels/hdg/mixedelasticitycreate.cpp>
+
+namespace Feel {
+namespace FeelModels {
+
+MIXEDELASTICITY_CLASS_INSTANTIATION
+
+}
+}
