+++ toolboxes/feel/feelmodels/hdg/mixedpoissonassembly_inst.cpp
@@ -1,0 +1,11 @@
+
+#include "mixedpoissonconfig.h"
+#include <feel/feelmodels/hdg/mixedpoissonassembly.cpp>
+
+namespace Feel {
+namespace FeelModels {
+
+MIXEDPOISSON_CLASS_INSTANTIATION
+
+}
+}
