+++ toolboxes/feel/feelmodels/hdg/mixedpoissonconfig.h.in
@@ -1,0 +1,20 @@
+#undef MIXEDPOISSON_DIM
+#undef MIXEDPOISSON_ORDERGEO
+#undef MIXEDPOISSON_ORDERPOLY
+
+#define MIXEDPOISSON_DIM @MIXEDPOISSON_DIM@
+#define MIXEDPOISSON_ORDERGEO @MIXEDPOISSON_ORDERGEO@
+#define MIXEDPOISSON_ORDERPOLY @MIXEDPOISSON_ORDERPOLY@
+
+#define MIXEDPOISSON_CLASS_TEMPLATE_DECLARATIONS \
+    template<int Dim, int Order, int G_Order, int E_Order> \
+        /**/
+#define MIXEDPOISSON_CLASS_TEMPLATE_TYPE \
+    MixedPoisson<Dim, Order, G_Order, E_Order> \
+        /**/
+#define MIXEDPOISSON_CLASS_TYPE \
+    MixedPoisson< MIXEDPOISSON_DIM,MIXEDPOISSON_ORDERPOLY,MIXEDPOISSON_ORDERGEO> \
+    /**/
+#define MIXEDPOISSON_CLASS_INSTANTIATION \
+    template class MIXEDPOISSON_CLASS_TYPE ;  \
+    /**/
