+++ toolboxes/feel/feelmodels/hdg/mixedpoissoncreate_inst.cpp
@@ -1,0 +1,11 @@
+
+#include "mixedpoissonconfig.h"
+#include <feel/feelmodels/hdg/mixedpoissoncreate.cpp>
+
+namespace Feel {
+namespace FeelModels {
+
+MIXEDPOISSON_CLASS_INSTANTIATION
+
+}
+}
