+++ toolboxes/hdg/CMakeLists.txt
@@ -28,34 +28,34 @@ if(CMAKE_C_COMPILER_ID STREQUAL "GNU" AND CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
 endif()
 
 foreach(D 2 3)
-    foreach(P RANGE 0 4 0) #3 
+    foreach(P RANGE 1 3 0) #3 
         feelpp_add_application( mixed-poisson-model_${D}DP${P}
                                 SRCS mixedpoisson_main.cpp
-                                LINK_LIBRARIES feelpp_modelcore
+                                LINK_LIBRARIES feelpp_toolbox_hdg_mixedpoisson_lib
                                 DEFS FEELPP_DIM=${D} FEELPP_ORDER=${P}
                                 INSTALL )
   endforeach()
 endforeach()
 
 foreach(D 3)
-  foreach(P RANGE 0 4 0) #3 
+  foreach(P RANGE 1 3 0) #3 
 	foreach (N 1)
 	  feelpp_add_application( mixed-elasticity-model_${D}DP${P}_G${N} 
 							SRCS mixedelasticity_main.cpp 
 							DEFS FEELPP_DIM=${D} FEELPP_ORDER=${P} FEELPP_GEO_ORDER=${N}
 							INCLUDE_IN_ALL
-							LINK_LIBRARIES feelpp_modelcore
+							LINK_LIBRARIES feelpp_toolbox_hdg_mixedelasticity_lib
                             INSTALL )
 	endforeach()
   endforeach()
 endforeach()
 
 foreach(D 2 3)
-  foreach(P RANGE 0 4 0) #3 
+  foreach(P RANGE 1 3) #3 
     feelpp_add_application( thermoelectricHDG_${D}DP${P}
                             SRCS thermoelectric.hpp thermoelectric_main.cpp
                             DEFS FEELPP_DIM=${D} FEELPP_ORDER=${P}
-                            LINK_LIBRARIES feelpp_modelcore
+                            LINK_LIBRARIES feelpp_toolbox_hdg_mixedpoisson_lib
                             INSTALL
                             )
   endforeach()
@@ -66,7 +66,7 @@ foreach(D 3)
     feelpp_add_application( poroelasticHDG_${D}DP${P}
                             SRCS poroelastic_main.cpp
                             DEFS FEELPP_DIM=${D} FEELPP_ORDER=${P}
-                            LINK_LIBRARIES feelpp_modelcore
+                            LINK_LIBRARIES feelpp_toolbox_hdg_mixedpoisson_lib feelpp_toolbox_hdg_mixedelasticity_lib
                             INSTALL
                             )
   endforeach()
@@ -80,3 +80,4 @@ feelpp_add_testcase(testDarcy CATEGORY hdg )
 feelpp_add_testcase(mixed-poisson CATEGORY hdg )
 feelpp_add_testcase(mixed-elasticity CATEGORY hdg )
 feelpp_add_testcase(mp-red CATEGORY hdg )
+feelpp_add_testcase(HL-31)
